changed directory structure(kevin is bitchy about it)
