+++ speaker_diarization/pyAudioAnalysis/roboy.gif
