+++ speaker_diarization/pyAudioAnalysis/roboy.pgm
