+++ speaker_diarization/pyAudioAnalysis/test.png
